Tarea Completada
Funcionalidad de productos completada

+++ ProyectoCapas/Entidades/Productos.cs
@@ -8,12 +8,16 @@ namespace Entidades
 {
     public class Productos
     {
+        /// <summary>
+        /// La clase Productos representa a la entidad productos de la base de datos
+        /// Los atributos de la clase representan a los campos de la entidad Productos
+        /// </summary>
         public int ProductID { get; set; }
         public string ProductName { get; set; }
         public int SupplierID { get; set; }
         public int CategoryID { get; set; }
         public int QuantityPerUnit { get; set; }
-        public int UnitPrice { get; set; }
+        public double UnitPrice { get; set; }
         public int UnitsInStock { get; set; }
         public int UnitsonOrder { get; set; }
         public int RecorderLevel { get; set; }
